+++ assets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.0.0+9]
+
+### Added
+
+- Firebase Crashlytics
+- Firebase Messaging
+
 ## [1.0.0+7]
 
 ### Added